requirements.txt
streamlit 
google-generativeai 
pillow
SpeechRecognition
textblob
spacy